requirements.txt
alembic==1.4.3
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
livereload==2.6.1
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==8.0.0
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.20
tornado==6.0.4
Werkzeug==1.0.0
WTForms==2.3.3
